-Wsign-conversion 警告を理由コメント付きキャストで解消する
タイムアウト ms と送信フラグなど、範囲が自明な符号変換を明示する。

+++ prod/libsrc/porter/api/potrCloseService.c
@@ -302,7 +302,9 @@ int potrCloseService(PotrContext *handle)
                 }
                 else
                 {
-                    close_result = wait_for_tcp_close_ack(ctx, ctx->global.tcp_close_timeout_ms);
+                    /* タイムアウト ms は設定値。実用範囲は INT_MAX 以下 */
+                    close_result =
+                        wait_for_tcp_close_ack(ctx, (int)ctx->global.tcp_close_timeout_ms);
                     if (close_result != POTR_OK)
                     {
                         POTR_TRACE(COM_UTIL_TRACE_LEVEL_WARNING,
