+++ prod/src/cmd/porter-test/porter-test.c
@@ -655,7 +655,8 @@ static void apply_send_command(PotrContext *handle, int is_file, char *cursor)
     {
         send_flags = 0;
     }
-    send_rtc = potrSend(handle, POTR_PEER_NA, send_data, send_len, send_flags | POTR_SEND_BLOCKING);
+    /* send_flags と POTR_SEND_* は 0x0003U 以下であり int に収まる */
+    send_rtc = potrSend(handle, POTR_PEER_NA, send_data, send_len, (int)(send_flags | POTR_SEND_BLOCKING));
     if (send_rtc != POTR_OK)
     {
         if (send_rtc == POTR_ERR_DISCONNECTED)
